requirements.txt
accelerate==1.12.0
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
exceptiongroup==1.3.1
filelock==3.20.2
fsspec==2025.12.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==25.0
portalocker==3.2.0
protobuf==6.33.2
psutil==7.2.1
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
qdrant-client==1.16.2
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.6.0
tqdm==4.67.1
transformers==4.57.3
triton==3.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
# 👇 [필수 추가됨] Hybrid Search & 형태소 분석용
rank-bm25
kiwipiepy